+++ PrayerOfTheDay/Strings.swift
@@ -10,10 +10,4 @@ import Foundation
 
 struct Strings {
     
-    static let tappedFacebookNotification = "com.prayerOfTheDay.notifications.tappedFacebook"
-    static let tappedGoogleNotification = "com.prayerOfTheDay.notifications.tappedGoogle"
-    static let tappedPinterestNotification = "com.prayerOfTheDay.notifications.tappedPinterest"
-    static let tappedTumblrNotification = "com.prayerOfTheDay.notifications.tappedTumblr"
-    static let tappedTwitterNotification = "com.prayerOfTheDay.notifications.tappedTwitter"
-    static let tappedYoutubeNotification = "com.prayerOfTheDay.notifications.tappedYoutube"
 }
